% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example dataset}
\format{A list of three matrices: each contains FRAP data for a control or experimental group. For each matrix, nrow = time_points + 1, ncol = sample size.}
\usage{
example_dataset
}
\description{
Example dataset
}
\keyword{datasets}
