% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Exclude samples from the dataset}
\usage{
exclude(ds, group, cols)
}
\arguments{
\item{ds}{Name of the dataset.}

\item{group}{Name of the group from which to exclude certain samples.}

\item{cols}{A vector of numbers specifying the column(s) to exclude.}
}
\value{
Modified dataset in the same format.
}
\description{
If certain samples are of poor quality, use this function to exclude them from the dataset.
}
\examples{
# after load("data/example_dataset.rda")
ds <- exclude(example_dataset, group = "mut1", cols = c(1,3))

}
