% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frapprocess.R
\name{frapprocess}
\alias{frapprocess}
\title{Process FRAP data}
\usage{
frapprocess(rawdata_dir, time_points = seq(0, 145, 5))
}
\arguments{
\item{rawdata_dir}{Absolute or relative path of the rawdata directory.}

\item{time_points}{A vector of time points (in second) that match raw data.}
}
\value{
The function has no return value. Instead, it generates output files:
\itemize{
\item One file for each group that contains normalized data.
\item One summary file of the sample means.
\item One summary file of the sample standard deviations.
\item One summary file of the parameters/results: ymax, ymin, k, halftime, tau, total_recovery, total_recovery_sd.
}
}
\description{
Takes quantified data from ImageJ, perform calculations and non-linear curve fitting. The raw data is organized as below:
\itemize{
\item There is a main folder
\item Each group (control or experimental) is in a subfolder
\item In each subfolder, there are n files for individual samples where n = sample size
}
}
\seealso{
Find more about FRAP, input data formats and methods used for the processing at the \href{https://github.com/GuanqiaoDing}{Github page}.
}
