% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frapplot.R
\name{frapplot}
\alias{frapplot}
\title{Plot FRAP data of two selected groups}
\usage{
frapplot(path, control, mutant, info)
}
\arguments{
\item{path}{Path of the output directory}

\item{control}{Name of the control.}

\item{mutant}{Name of the mutant.}

\item{info}{Returned information from \code{\link[=frapprocess]{frapprocess()}}.}
}
\description{
Plot FRAP data of any two groups (e.g. control and mutant) in a consistent and publishable format.
}
\examples{
info <- frapprocess(example_dataset, seq(0, 145, 5))
frapplot(tempdir(), "control", "mut2", info)

}
