% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frapplot.R
\name{frapplot}
\alias{frapplot}
\title{Plot FRAP data of two samples}
\usage{
frapplot(control, mutant)
}
\arguments{
\item{control}{Name of the control.}

\item{mutant}{Name of the mutant.}
}
\value{
The function has no return value. Instead, it outputs the plot as a pdf file.
}
\description{
Takes in group names (e.g. control, mutant, etc.) and plot their data. Note: \code{\link[=frapprocess]{frapprocess()}} needs to run and process the data first.
}
\seealso{
Find more about FRAP, input data formats and methods used for the processing at the \href{https://github.com/GuanqiaoDing}{Github page}.
}
