% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frapprocess.R
\name{frapprocess}
\alias{frapprocess}
\title{Process FRAP data}
\usage{
frapprocess(ds, time_points)
}
\arguments{
\item{ds}{A dataset that contains FRAP data for multiple experiment groups}

\item{time_points}{A vector of time points (in second) that the experiment uses, e.g. 0, 5, 10, ....}
}
\value{
A list of results:
\itemize{
\item $time_points: a vector of time points
\item $summary: summary of the analysis
\item $sample_means: a matrix of sample means, nrow = num of time points, ncol = sample size
\item $sample_sd: a matrix of standard deviations, nrow = num of time points, ncol = sample size
\item $model: result of non-linear regression model
}
}
\description{
Normalize and analyze FRAP data. Perform non-linear regression and calculate ymax, ymin, k, halftime, tau, total_recovery, total_recovery_sd.

Reference: Brian L. Sprague, et al. Analysis of Binding Reactions by Fluorescence Recovery after Photobleaching. Biophys J (2004) \url{https://doi.org/10.1529/biophysj.103.026765}
}
\examples{
# after load("data/example_dataset.rda")
info <- frapprocess(example_dataset, seq(0, 145, 5))

}
