% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frapprocess.R
\name{frapprocess}
\alias{frapprocess}
\title{Process FRAP data}
\usage{
frapprocess(rawdata_dir, time_points = seq(0, 145, 5))
}
\arguments{
\item{rawdata_dir}{Absolute or relative path of the rawdata directory.}

\item{time_points}{A vector of time points (in second) that match raw data.}
}
\value{
Results used for the plot as a list.
It also generates output files:
\itemize{
\item One file for each group that contains normalized data.
\item One summary file of the sample means.
\item One summary file of the sample standard deviations.
\item One summary file of the parameters/results: ymax, ymin, k, halftime, tau, total_recovery, total_recovery_sd.
}
}
\description{
Takes time-lapse data from ImageJ after the intensity quantification, perform normalization, calculations and non-linear curve fitting. The results are written to output files and returned to be used for \code{\link[=frapplot]{frapplot()}}.
The raw data is organized as below, which is complied with the regular practice when processing imageJ time-lapse data by hand:
\itemize{
\item There is a main folder for the experiment (rawdata direcotry)
\item Each group (e.g. control, mutant1, mutant2, etc.) is in a subfolder
\item In each subfolder, there are multiple ".txt" files for individual samples where the number equals to the sample size (the function uses ".txt" as a pattern to read all relavant files)
\item In each file, the first value is the fluorecent intensity before the bleach, followed by the time-lapse values, one value per line
}
}
\examples{
\dontrun{
info <- frapprocess("~/experiment/rawdata")
frapplot("name1", "name2", info)
}
}
